fix: make errorHook a first-class citizen in I/O handling
The current implementation would not do anything when in runtime error states.
This commit adds the processing of driver states (error, timeout) and introduces
the errorProcessIO handler that can be overriden. By default, the handler calls
processIO.

To keep the current behaviour, all of this needs to be explicitly enabled by the
subclasses by calling setEnableRuntimeErrorProcessing()